requirements.txt
numpy==1.23.4
opencv-python==4.6.0.66
pyinstaller==5.6.2
PyYAML==6.0
tqdm==4.64.1
